fix(mpnn_message): the driver built a bf16 mask the contract requires in fp32
Four of the new drivers failed on the card with "requires BF16 projection math". Not the projection:
`interface._triton_contract_supported` requires `edge_mask.dtype == torch.float32`, and mine built it
at the activation's dtype. The reduction multiplies by the mask in fp32, so a bf16 one rounds the
weight before the sum rather than after -- the driver builds it at the dtype the kernel is entitled
to now, with that written where it is built.

Card, 22 kernels driven: 15 matched a reference, 7 open. All eight of the newly registered ones pass
-- mpnn_message 1.04e-03 forward and 9.60e-03 through its gradients, mpnn_edge_mlp 4.28e-03 on both
policies, and the dropout pack at 0.00e+00, which is what a bit-packing should give.

The seven that remain are the seven that were already open, unchanged by any of this work: five
outside a band nobody has declared, two that will not launch on an untuned card.

Both pages regenerated from the registry and from the harness log rather than transcribed --
93 kernels on the sweep page, 22 rows on the mpnn one with each pill read out of the run.

+++ src/miniworld_engine/kernels/checks/mpnn_message.py
@@ -22,7 +22,7 @@ def _pair():
         out = message_hidden_reduce(
             preactivation, weight, bias, mask, _NEIGHBORS, backend="triton")
     ref = message_hidden_reduce_pytorch(
-        preactivation.float(), weight.float(), bias.float(), mask.float(), _NEIGHBORS)
+        preactivation.float(), weight.float(), bias.float(), mask, _NEIGHBORS)
     return out, ref
 
 
@@ -41,7 +41,7 @@ def _gradients():
             return message_hidden_reduce(*a, mask, _NEIGHBORS, backend="triton")
 
     def reference(*a):
-        return message_hidden_reduce_pytorch(*a, mask.float(), _NEIGHBORS)
+        return message_hidden_reduce_pytorch(*a, mask, _NEIGHBORS)
 
     return _grads(kernel, [preactivation, weight, bias], reference, names)
